+++ Horizontes_Urbanos/src/main/java/Modelo/entities/Administrador.java
@@ -18,7 +18,7 @@ import javax.persistence.Table;
 public class Administrador implements Serializable {
     
     @Id
-    private int cedula;
+    private Long cedula;
     private String nombre;
     private String direccion;
     
@@ -28,18 +28,18 @@ public class Administrador implements Serializable {
     public Administrador() {
     }
 
-    public Administrador(int cedula, String nombre, String direccion, ArrayList<Proyecto> listaProyectos) {
+    public Administrador(Long cedula, String nombre, String direccion, ArrayList<Proyecto> listaProyectos) {
         this.cedula = cedula;
         this.nombre = nombre;
         this.direccion = direccion;
         this.listaProyectos = listaProyectos;
     }
 
-    public int getCedula() {
+    public Long getCedula() {
         return cedula;
     }
 
-    public void setCedula(int cedula) {
+    public void setCedula(Long cedula) {
         this.cedula = cedula;
     }
