+++ Horizontes_Urbanos/src/main/java/Modelo/jpa_controllers/PagoJpaController.java
@@ -132,7 +132,7 @@ public class PagoJpaController implements Serializable {
         } catch (Exception ex) {
             String msg = ex.getLocalizedMessage();
             if (msg == null || msg.length() == 0) {
-                int id = pago.getId_pago();
+                Long id = pago.getId_pago();
                 if (findPago(id) == null) {
                     throw new NonexistentEntityException("The pago with id " + id + " no longer exists.");
                 }
@@ -145,7 +145,7 @@ public class PagoJpaController implements Serializable {
         }
     }
 
-    public void destroy(int id) throws NonexistentEntityException {
+    public void destroy(Long id) throws NonexistentEntityException {
         EntityManager em = null;
         try {
             em = getEntityManager();
@@ -205,7 +205,7 @@ public class PagoJpaController implements Serializable {
         }
     }
 
-    public Pago findPago(int id) {
+    public Pago findPago(Long id) {
         EntityManager em = getEntityManager();
         try {
             return em.find(Pago.class, id);
