+++ Horizontes_Urbanos/src/main/java/Modelo/persistir/IPersistenciaTorre.java
@@ -9,5 +9,5 @@ import Modelo.entities.Torre;
  */
 public interface IPersistenciaTorre extends IPersistencia<Torre> {
     
-    int contarNAptos(int numero_torre,int id_proyecto);
+    int contarNAptos(int numero_torre,Long id_proyecto);
 }
